Fix David Reiser's compile problem on OSX with gcc-3.3.
	* src/core-utils/gnc-gkeyfile-utils.c: cast the
	  printf so we don't warn when sizeof(ssize_t) !=
	  sizeof(int).  In this case the cast is safe because
	  we can only write an int-size of data.



git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@11925 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/core-utils/gnc-gkeyfile-utils.c
@@ -123,10 +123,10 @@ gnc_key_file_save_to_file (const gchar *filename,
     if (error) {
       *error = g_error_new(G_FILE_ERROR, g_file_error_from_errno(errno), 
 			   "File %s truncated (provided %d, written %d)",
-			   filename, length, written);
+			   filename, length, (int) written);
     } else {
       g_critical("File %s truncated (provided %d, written %d)",
-	      filename, length, written);
+	      filename, length, (int) written);
     }
     /* Ignore any error */
     close(fd);
